Move Menu to Future structure
