+++ .changeset/violet-cheetahs-kneel.md
@@ -1,0 +1,5 @@
+---
+"@kaizen/components": minor
+---
+
+Move Menu v3 to Future
