+++ packages/components/src/Filter/FilterBar/subcomponents/AddFiltersMenu/AddFiltersMenu.tsx
@@ -1,7 +1,8 @@
 import React from "react"
 import { useIntl } from "@cultureamp/i18n-react-intl"
 import { AddIcon } from "~components/Icon"
-import { Menu, MenuList, MenuItem, Button } from "~components/__actions__/v2"
+import { Menu, MenuList, MenuItem } from "~components/Menu"
+import { Button } from "~components/__actions__/v2"
 import { useFilterBarContext } from "../../context/FilterBarContext"
 
 export const AddFiltersMenu = (): JSX.Element => {
