+++ packages/components/src/Menu/_docs/Menu.stickersheet.stories.tsx
@@ -10,7 +10,7 @@ import { Menu } from "../index"
 import { MenuContentExample } from "./MenuContentExample"
 
 export default {
-  title: "Actions/Menu/v1",
+  title: "Actions/Menu",
   parameters: {
     chromatic: { disable: false },
     controls: { disable: true },
