+++ packages/components/src/Menu/_docs/Menu.stories.tsx
@@ -2,11 +2,13 @@ import React from "react"
 import { Meta, StoryObj } from "@storybook/react"
 import { ChevronDownIcon, DuplicateIcon } from "~components/Icon"
 import { Button } from "~components/__actions__/v2"
-import * as MenuV1Stories from "../../v1/_docs/Menu.stories"
-import { Menu, MenuList, MenuItem, MenuHeading } from "../index"
+import { Menu } from "../index"
+import { MenuHeading } from "../subcomponents/MenuHeading"
+import { MenuItem } from "../subcomponents/MenuItem"
+import { MenuList } from "../subcomponents/MenuList"
 
 const meta = {
-  title: "Actions/Menu/v2",
+  title: "Actions/Menu",
   component: Menu,
   args: {
     button: (
@@ -29,21 +31,14 @@ const meta = {
           <MenuItem
             href="https://cultureamp.atlassian.net/wiki/spaces/DesignSystem/pages/3082059782/Menu"
             label="Learn more about Menu"
-            destructive
           />
         </MenuList>
       </MenuList>
     ),
   },
-  decorators: [
-    Story => (
-      <div className="flex mt-[60px] gap-12">
-        <Story />
-      </div>
-    ),
-  ],
-  parameters: {
-    layout: "centered",
+  argTypes: {
+    children: { control: false },
+    button: { control: false },
   },
 } satisfies Meta<typeof Menu>
 
@@ -51,4 +46,12 @@ export default meta
 
 type Story = StoryObj<typeof meta>
 
-export const Playground: Story = MenuV1Stories.Playground
+export const Playground: Story = {
+  parameters: {
+    docs: {
+      canvas: {
+        sourceState: "shown",
+      },
+    },
+  },
+}
