+++ packages/components/src/TitleBlockZen/subcomponents/MobileActions.tsx
@@ -2,12 +2,8 @@ import React, { useCallback, useEffect, useState } from "react"
 import classnames from "classnames"
 import { FocusOn } from "react-focus-on"
 import { ChevronDownIcon, ChevronUpIcon } from "~components/Icon"
-import {
-  MenuItem,
-  MenuList,
-  MenuHeading,
-  ButtonProps,
-} from "~components/__actions__/v2"
+import { MenuItem, MenuList, MenuHeading } from "~components/Menu"
+import { ButtonProps } from "~components/__actions__/v2"
 import { TITLE_BLOCK_ZEN_OTHER_ACTIONS_HTML_ID } from "../constants"
 import {
   DefaultActionProps,
