+++ packages/components/src/TitleBlockZen/subcomponents/SecondaryActions.tsx
@@ -1,6 +1,7 @@
 import React from "react"
 import { ChevronDownIcon, MeatballsIcon } from "~components/Icon"
-import { Menu, MenuList, Button, IconButton } from "~components/__actions__/v2"
+import { Menu, MenuList } from "~components/Menu"
+import { Button, IconButton } from "~components/__actions__/v2"
 import styles from "../TitleBlockZen.module.scss"
 import { TITLE_BLOCK_ZEN_SECONDARY_MENU_HTML_ID } from "../constants"
 import { SecondaryActionsProps, TitleBlockMenuItemProps } from "../types"
