+++ packages/components/src/TitleBlockZen/subcomponents/Toolbar.tsx
@@ -4,7 +4,8 @@
 
 import React from "react"
 import classnames from "classnames"
-import { MenuProps, ButtonProps } from "~components/__actions__/v2"
+import { MenuProps } from "~components/Menu"
+import { ButtonProps } from "~components/__actions__/v2"
 import styles from "./Toolbar.module.scss"
 
 type ToolbarProps = {
