+++ packages/components/src/__actions__/v1.ts
@@ -1,2 +1,1 @@
-export * from "./Menu/v1"
 export * from "./Button/v1"
