+++ packages/components/src/__actions__/v2.ts
@@ -1,2 +1,1 @@
-export * from "./Menu/v2"
 export * from "./Button/v2"
