+++ packages/components/src/__future__/Menu/_docs/ApiSpecification.mdx
@@ -3,12 +3,12 @@ import { KAIOInstallation, ResourceLinks,  SbContent } from "~storybook/componen
 import * as docsStories from "./Menu.docs.stories"
 import * as exampleStories from "./Menu.stories"
 
-<Meta title="Actions/Menu/v3/API Specification"/>
+<Meta title="Actions/Menu/Future/API Specification"/>
 
 <SbContent>
-  # Menu API Specification (v3)
+  # Menu API Specification (Future)
 
-  Updated July 4, 2024
+  Updated Sep 9, 2024
 </SbContent>
 
 <ResourceLinks
@@ -22,15 +22,15 @@ import * as exampleStories from "./Menu.stories"
   <div className="bg-yellow-100 border-default border-yellow-400 rounded px-24">
     ### Disclaimer
 
-    The `Menu` `v3` component is not backwards compatible with Kaizen `Button` `v1` and `v2`. For this reason, usage of this component is not recommended until a `Button` `v3` is released.
+    The Future `Menu` component is not backwards compatible with the current Kaizen `Button`. For this reason, usage of this component is not recommended until a Future `Button` is released.
   </div>
 </SbContent>
 
 <SbContent className="mb-24">
 
-  <KAIOInstallation exportNames={["Menu", "MenuTrigger", "MenuItem"]} family="actions" version="3" />
+  <KAIOInstallation exportNames={["Menu", "MenuTrigger", "MenuItem"]}/>
 
-  <Source code={"import { Popover, Section, Header } from \"@kaizen/components/v3/react-aria-components\""} language="tsx" />
+  <Source code={"import { Popover, Section, Header } from \"react-aria-components\""} language="tsx" />
 
   ## Overview
