+++ packages/components/src/__future__/Menu/_docs/Menu.mdx
@@ -3,10 +3,10 @@ import { ResourceLinks, SbContent, Installation, DosAndDonts, DoOrDont } from "~
 import * as MenuDocsStories from "./Menu.docs.stories"
 import * as MenuStories from "./Menu.stories"
 
-<Meta title="Actions/Menu/v3/Usage Guidelines" />
+<Meta title="Actions/Menu/Future/Usage Guidelines" />
 
 <SbContent>
-  # Menu (v3)
+  # Menu (future)
 
   Updated July 6, 2024
 </SbContent>
