+++ packages/components/src/index.ts
@@ -30,6 +30,7 @@ export * from "./Label"
 export * from "./LabelledMessage"
 export * from "./LikertScaleLegacy"
 export * from "./Loading"
+export * from "./Menu"
 export * from "./Modal"
 export * from "./MultiSelect"
 export * from "./Notification"
@@ -59,5 +60,4 @@ export * from "./VisuallyHidden"
 export * from "./Well"
 export * from "./__layout__/Workflow/v1"
 export * from "./__overlays__/Tooltip/v1"
-export * from "./__actions__/Menu/v1"
 export * from "./__actions__/Button/v1"
